feat: integrate measurements API and refactor measurements handling
- Updated MeasurementsRepo to utilize ApiService for real API calls instead of mock data.
- Added new API endpoints for fetching, updating, and deleting measurements.
- Refactored MeasurementsCubit to handle API responses and manage state accordingly.
- Removed deprecated MeasurementsModel and adjusted related UI components to work with new data structures.
- Enhanced measurement input handling to support integer values and improved form validation.

+++ lib/features/customer/measurements/ui/widgets/measurement_field.dart
@@ -47,7 +47,7 @@ class MeasurementField extends StatelessWidget {
         if (value == null || value.trim().isEmpty) {
           return 'Required';
         }
-        final parsed = double.tryParse(value.trim());
+        final parsed = int.tryParse(value.trim());
         if (parsed == null || parsed <= 0) {
           return 'Enter a valid number';
         }
